Небольшие изменения. 1. Добавлено API для работы с таблицей School. 2. Сканер штрих-кодов перемещён в Android-приложение. 3. Рефакторинг и починка работы с Excel. 4. Обновление документации (комментарии и README).

+++ backend/database/result.py
@@ -9,7 +9,7 @@ class Result(SqlAlchemyBase, Table):
     student_code = sa.Column(sa.Integer, nullable=False, primary_key=True)
     student_id = sa.Column(sa.Integer, nullable=False)
     result = sa.Column(sa.Float, nullable=False)
-    net_score = sa.Column(sa.Integer, nullable=False)
+    net_score = sa.Column(sa.Float, nullable=False)
     position = sa.Column(sa.Integer, nullable=False)
 
     @staticmethod
